feat: ebpf

+++ pkg/core/hooks/kernelComm.go
@@ -66,7 +66,7 @@ func (h *Hooks) CleanProxyEntry(srcPort uint16) error {
 
 func (h *Hooks) SendAppInfo(appInfo structs.AppInfo) error {
 	key := 1234
-	err := h.appRegistrationMap.Update(uint32(key), appInfo, ebpf.UpdateAny)
+	err := h.appRegistrationMap.Update(uint64(key), appInfo, ebpf.UpdateAny)
 	if err != nil {
 		utils.LogError(h.logger, err, "failed to send the app info to the ebpf program")
 		return err
